# services/orders/openapi.yaml
openapi: 3.0.3

info:
  title: CampusEats Orders Service
  version: "1.0.0"
  description: REST API for managing CampusEats orders.

security:
  - bearerAuth: []

servers:
  - url: http://localhost:8081

paths:

  /orders:
    post:
      summary: Create an order
      operationId: createOrder

      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"

      responses:
        "201":
          description: Order created
          headers:
            Location:
              description: URL of the newly created order
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          $ref: "#/components/responses/BadRequest"

        "409":
          $ref: "#/components/responses/Conflict"

        "422":
          $ref: "#/components/responses/UnprocessableEntity"

        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    get:
      summary: List orders for a student
      operationId: listOrders

      parameters:
        - in: query
          name: studentId
          required: true
          description: Student whose orders should be returned.
          schema:
            type: integer
            minimum: 1

      responses:
        "200":
          description: Orders returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

        "400":
          $ref: "#/components/responses/BadRequest"


  /orders/{id}:
    get:
      summary: Get one order
      operationId: getOrder

      parameters:
        - $ref: "#/components/parameters/OrderId"

      responses:
        "200":
          description: Order returned
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Replace an order representation using optimistic concurrency
      operationId: replaceOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Order replaced
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"


  /orders/{id}/cancellation:
    post:
      summary: Cancel an order
      operationId: cancelOrder

      parameters:
        - $ref: "#/components/parameters/OrderId"

      responses:
        "202":
          description: Cancellation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "404":
          $ref: "#/components/responses/NotFound"

        "409":
          $ref: "#/components/responses/Conflict"


  /carts/items:
    post:
      summary: Add an item to a cart
      operationId: addCartItem

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"

      responses:
        "201":
          description: Cart item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

        "409":
          $ref: "#/components/responses/Conflict"


components:

  parameters:

    OrderId:
      in: path
      name: id
      required: true
      description: Unique order identifier.
      schema:
        type: integer
        minimum: 1

    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Unique key used to safely retry order creation.
      schema:
        type: string
        minLength: 1

    IfMatch:
      in: header
      name: If-Match
      required: true
      description: ETag read from the current order representation.
      schema:
        type: string


  responses:

    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    Conflict:
      description: Request conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    UnprocessableEntity:
      description: Request is syntactically valid but rejected by the domain.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    ServiceUnavailable:
      description: Required downstream service is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    PreconditionFailed:
      description: The supplied If-Match ETag is stale.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"

    PreconditionRequired:
      description: An update requires an If-Match header.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"


  schemas:

    CreateOrderRequest:
      type: object
      required:
        - studentId
        - items
      properties:
        studentId:
          type: integer
          minimum: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"

    AddCartItemRequest:
      type: object
      required:
        - studentId
        - itemId
        - quantity
      properties:
        studentId:
          type: integer
          minimum: 1
        itemId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1

    OrderItem:
      type: object
      required:
        - itemId
        - quantity
      properties:
        itemId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1

    Order:
      type: object
      required:
        - id
        - studentId
        - items
        - status
      properties:
        id:
          type: integer
          minimum: 1
        studentId:
          type: integer
          minimum: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum:
            - placed
            - cancelled

    Cart:
      type: object
      required:
        - studentId
        - items
      properties:
        studentId:
          type: integer
          minimum: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token